* Now parsers a migration plan and executes it via the Indexer :)

+++ tools/plugins/formatrisks/trunk/archive_name/cfg/plugins/EPrints/Plugin/Screen/Admin/FormatsRisks_delete_plan.pm
@@ -40,14 +40,19 @@ sub action_delete_plan {
 
 		my $list = $searchexp->perform_search;
 		my $failed_flag = 0;
+		my $msg = "success";
 		$list->map( sub {
 			my $preservation_plan = $_[2];
-			my $file_path = $preservation_plan->get_value("file_path");
-			unlink($file_path);
-			if (!-s "$file_path") {
-				$preservation_plan->remove();
+			if ($session->in_use($preservation_plan)<1){		
+				my $file_path = $preservation_plan->get_value("file_path");
+				unlink($file_path);
+				if (!-s "$file_path") {
+					$preservation_plan->remove();
+				} else {
+					$msg = "failed_in_delete";
+				}
 			} else {
-				$failed_flag = 1;
+				$msg = "in_use";
 			}
 		});
 
@@ -69,6 +74,30 @@ sub action_delete_plan {
 				$self->html_phrase( "Failed" )
 				);
 		$self->{processor}->{screenid} = "Admin::FormatsRisks";
+	}
+
+}
+
+sub in_use 
+{
+	my ( $session, $preservation_plan ) = @_;
+	
+	my $dataset = $session->get_repository->get_dataset( "document" );
+	
+	my $pres_plan_uri = $preservation_plan->internal_uri();
+	my $searchexp = EPrints::Search->new(
+			session => $session,
+			dataset => $dataset,
+			filters => [
+			{ meta_fields => [qw( relation_uri )], value => "$pres_plan_uri", match => "EX" },
+			],
+			);
+	
+	my $list = $searchexp->perform_search;
+	if ($list->count() > 0) {
+		return 1;
+	} else {
+		return 0;
 	}
 
 }
